+++ src/main/java/com/example/bankoflykhvar/service/AccountService.java
@@ -1,12 +1,13 @@
 package com.example.bankoflykhvar.service;
 
 import com.example.bankoflykhvar.model.Account;
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface AccountService {
     Account add(Account account);
 
-    Double getBalance(String accountNumber);
+    BigDecimal getBalance(String accountNumber);
 
     List<Account> getAllUserAccountsByPhoneNumber(String phoneNumber);
